Added makefiles to all subfolders that were missing them and renamed a few files.
