+++ searches/binary-search/Makefile
@@ -1,7 +1,7 @@
 PROG=search
 CC=gcc
 CFLAGS=-std=c99
-SRC=main.c binary_search.c
+SRC=binary_search.c
 
 .PHONY: all
 
@@ -11,7 +11,7 @@ test: all
 	./${PROG}
 
 ${PROG}: ${SRC}
-	${CC} ${CFLAGS} ${SRC} -o ${PROG}
+	${CC} ${CFLAGS} ${SRC} -c
 	
 clean: 
 	@if test -f ${PROG}; then rm -f ${PROG}; fi
