+++ sorts/mergesort/Makefile
@@ -1,0 +1,16 @@
+
+CC=gcc
+CFLAGS=-std=c99
+SRC=mergesort.c
+
+.PHONY: all
+
+all: ${SRC}
+	${CC} ${CFLAGS} $^ -c
+
+# TODO add test
+#test: all
+#	./${PROG}
+	
+clean: 
+	@if test -f *.o; then rm -f *.o; fi
