+++ sorts/selection-sort/Makefile
@@ -7,8 +7,15 @@ SRC=test.c selectionSort.c
 
 all: ${PROG}
 
-${PROG}: ${SRC}
+test: all
+	./${PROG}
+
+${PROG}: test.c selectionSort.o
 	${CC} ${CFLAGS} ${SRC} -o ${PROG}
+
+build: selectionSort.c
+	${CC} ${CFLAGS} -c $^
 	
 clean: 
 	@if test -f ${PROG}; then rm -f ${PROG}; fi
+	@if test -f *.o; then rm -f *.o; fi
